[system-probe] Introduce module API (#5544)
